fix: change nonterminal count paramter type to uint16_t

+++ include/grammar.h
@@ -27,7 +27,7 @@ struct rdesc_grammar {
 	const struct rdesc_grammar_symbol *rules;
 
 	/** @brief Total number of nonterminals. */
-	uint32_t nt_count;
+	uint16_t nt_count;
 
 	/**
 	 * @brief Maximum number of variants, used to segment the production
@@ -75,7 +75,7 @@ extern "C" {
 
 /** @brief Initializes a grammar struct. */
 void rdesc_grammar_init(struct rdesc_grammar *grammar,
-			uint32_t nonterminal_count,
+			uint16_t nonterminal_count,
 			uint16_t nonterminal_variant_count,
 			uint16_t nonterminal_body_length,
 			const struct rdesc_grammar_symbol *production_rules);
